Updated some data type of fields in routeModel class.

+++ app/src/main/java/com/cmpt276/indigo/carbontracker/carbon_tracker_model/RouteModel.java
@@ -7,9 +7,9 @@ package com.cmpt276.indigo.carbontracker.carbon_tracker_model;
 public class RouteModel implements CarbonFootprintComponent{
     private long id;
     private String name;
-    private int cityDistance;
-    private int highwayDistance;
-    private int totalDistance;
+    private double cityDistance;
+    private double highwayDistance;
+    private double totalDistance;
     private boolean isDeleted;
 
     public RouteModel(){
@@ -20,7 +20,7 @@ public class RouteModel implements CarbonFootprintComponent{
         isDeleted = false;          // when a Route is deleted, we should hide it instead of removing it
     }
 
-    public RouteModel(long id, String name, int cityDistance, int highwayDistance, int totalDistance, boolean isDeleted){
+    public RouteModel(long id, String name, double cityDistance, double highwayDistance, double totalDistance, boolean isDeleted){
         this.id = id;
         this.name = name;
         this.cityDistance = cityDistance;
@@ -37,15 +37,15 @@ public class RouteModel implements CarbonFootprintComponent{
         return name;
     }
 
-    public int getCityDistance() {
+    public double getCityDistance() {
         return cityDistance;
     }
 
-    public int getHighwayDistance() {
+    public double getHighwayDistance() {
         return highwayDistance;
     }
 
-    public int getTotalDistance() {
+    public double getTotalDistance() {
         totalDistance = cityDistance + highwayDistance;
         return totalDistance;
     }
@@ -58,21 +58,21 @@ public class RouteModel implements CarbonFootprintComponent{
         this.name = name;
     }
 
-    public void setCityDistance(int cityDistance) {
+    public void setCityDistance(double cityDistance) {
         if(cityDistance < 0){
             throw new IllegalArgumentException("City distance must be a positive number.");
         }
         this.cityDistance = cityDistance;
     }
 
-    public void setHighwayDistance(int highwayDistance) {
+    public void setHighwayDistance(double highwayDistance) {
         if(highwayDistance < 0){
             throw new IllegalArgumentException("Highway distance must be a positive number.");
         }
         this.highwayDistance = highwayDistance;
     }
 
-    public void setTotalDistance(int totalDistance) {
+    public void setTotalDistance(double totalDistance) {
         if(totalDistance < 0){
             throw new IllegalArgumentException("Total distance must be a positive number.");
         }
